move ToJSON instances to type definitions
